refactor io_loop
